requirements.txt
django>=4.1.0,<4.2.0
asgiref==3.5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pyjwt==2.6.0
psycopg2-binary==2.9.5
requests==2.28.1
django-filter==22.1
sqlparse==0.4.3
pytz==2022.6
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
tomli==2.0.1
urllib3==1.26.13